+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Add retry logic to the client
+
 ## [0.2.0] - 2023-11-08
 
 - Wrap the response of OpenMeteo::Forecast to get back an OpenMeteo::Entities::Forecast entity