+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add minutely_15 to forecast variables
 - Add minutely_15 to forecast entity
 - Add DWD Icon forecast model
+- Add integration test with VCR
 
 ## [0.2.0] - 2023-11-08